+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptsnippetop.py
@@ -1846,7 +1846,7 @@ class CreatePromptSnippetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-05T07:31:09.342Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-05T11:00:10.120Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")