+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -106,5 +106,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-02-05T07:31:10.392Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-02-05T11:00:11.228Z")
     r"""The date and time the resource was created"""