+++ packages/orq-rc/RELEASES.md
@@ -129,3 +129,13 @@ Based on:
 - [python v3.1.0-rc.35] packages/orq-rc
 ### Releases
 - [PyPI v3.1.0-rc.35] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.35 - packages/orq-rc
+
+## 2025-02-05 11:01:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.484.1 (2.503.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.1.0-rc.36] packages/orq-rc
+### Releases
+- [PyPI v3.1.0-rc.36] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.36 - packages/orq-rc